requirements.txt
umap-learn==0.3.10
seaborn==0.9.0
pandas==0.24.2
numpy==1.16.4
matplotlib==3.1.0
